Decouple editor app itself from app providers
